requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.14,<4.0
Pillow>=10.0,<11.0
gunicorn>=21.0,<23.0
whitenoise>=6.6,<7.0
python-dotenv>=1.0,<2.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
cloudinary>=1.36,<2.0
django-cloudinary-storage>=0.3,<1.0
